refactor (PullRequest): refatorando retirando o registro local dos pullrequests
