--- app/src/main/java/com/testemobile/githubjava/Model/OwnerModelRepo.kt
@@ -1,14 +1,0 @@
-package com.testemobile.githubjava.Model
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-
-@Entity(tableName = "owner")
- class OwnerModelRepo {
-
-    @ColumnInfo(name="login")
-    var login: String=""
-
-    @ColumnInfo(name="avatar_url")
-    val avatarUrl: String=""
-}
