+++ app/src/main/java/com/testemobile/githubjava/Model/PullRequestModel.kt
@@ -4,7 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class PullRequestModel(
 
-
     @SerializedName("title")
     var tituloPullRequests: String,
