+++ app/src/main/java/com/testemobile/githubjava/View/Holder/ListHolder.kt
@@ -1,4 +1,4 @@
-package com.testemobile.githubjava.Holder
+package com.testemobile.githubjava.View.Holder
 
 import androidx.recyclerview.widget.RecyclerView
 import com.testemobile.githubjava.Model.ItemsModel
