+++ app/src/main/java/com/testemobile/githubjava/module/AppModule.kt
@@ -1,15 +1,12 @@
 package com.testemobile.githubjava.module
 
-import com.testemobile.githubjava.ViewModel.PullRequestViewModel
 import com.testemobile.githubjava.ViewModel.RepositorioViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
-import org.koin.core.context.loadKoinModules
 import org.koin.dsl.module
 
 val appModule = module {
 
     viewModel{RepositorioViewModel(get())}
-    viewModel{PullRequestViewModel(get())}
 }
